Move "for" command in a separate file.
